requirements.txt
requests==2.22.0
pycodestyle==2.5.0
Pillow==8.3.2
